+++ hosts/default/variables.nix
@@ -90,6 +90,6 @@
   #animChoice = ../../modules/home/hyprland/animations-moving.nix;
 
   # Set network hostId if required (needed for zfs)
-  # Otherwise leave as-is 
-  hostId = "5ab03f50;
+  # Otherwise leave as-is
+  hostId = "5ab03f50";
 }