+++ hosts/zaneyos-oem/variables.nix
@@ -75,4 +75,8 @@
   #animChoice = ../../modules/home/hyprland/animations-def.nix;
   #animChoice = ../../modules/home/hyprland/animations-dynamic.nix;
   #animChoice = ../../modules/home/hyprland/animations-moving.nix;
+
+  # Set network hostId if required (needed for zfs)
+  # Otherwise leave as-is
+  hostId = "5ab03f50";
 }